+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 74:
 * Add file_stdio and File concept
 * Add file_win32
 * Add file_body
+* Remove common/file_body.hpp
 
 --------------------------------------------------------------------------------